requirements.txt
Django==3.2.9
django-connection-url==0.1.2
django-extensions==3.1.5
ipython==7.30.1
psycopg2-binary==2.9.2
django-bootstrap4==21.1
whitenoise==5.3.0
django-import-export==2.6.1
